+++ codec/hessian/encode.go
@@ -27,8 +27,8 @@ import (
 // array object struct
 
 type Encoder struct {
-	clsDefList []classDef
-	buffer     []byte
+	classInfoList []classInfo
+	buffer        []byte
 }
 
 func NewEncoder() *Encoder {
@@ -62,6 +62,15 @@ func (e *Encoder) Encode(v interface{}) error {
 	case bool:
 		e.buffer = encBool(v.(bool), e.buffer)
 
+	case int8:
+		e.buffer = encInt32(v.(int32), e.buffer)
+
+	case int16:
+		e.buffer = encInt32(v.(int32), e.buffer)
+
+	case int32:
+		e.buffer = encInt32(v.(int32), e.buffer)
+
 	case int:
 		// if v.(int) >= -2147483648 && v.(int) <= 2147483647 {
 		// 	b = encInt32(int32(v.(int)), b)
@@ -70,9 +79,6 @@ func (e *Encoder) Encode(v interface{}) error {
 		// }
 		// 把int统一按照int64处理，这样才不会导致decode的时候出现" reflect: Call using int32 as type int64 [recovered]"这种panic
 		e.buffer = encInt64(int64(v.(int)), e.buffer)
-
-	case int32:
-		e.buffer = encInt32(v.(int32), e.buffer)
 
 	case int64:
 		e.buffer = encInt64(v.(int64), e.buffer)
@@ -311,10 +317,10 @@ func encString(v string, b []byte) []byte {
 /////////////////////////////////////////
 
 // # 8-bit binary data split into 64k chunks
-// ::= x41 b1 b0 <binary-data> binary # non-final chunk
-// ::= 'B' b1 b0 <binary-data>        # final chunk
-// ::= [x20-x2f] <binary-data>        # binary data of length 0-15
-// ::= [x34-x37] <binary-data>        # binary data of length 0-1023
+// ::= x41(A) b1 b0 <binary-data> binary # non-final chunk
+// ::= x42(B) b1 b0 <binary-data>        # final chunk
+// ::= [x20-x2f] <binary-data>           # binary data of length 0-15
+// ::= [x34-x37] <binary-data>           # binary data of length 0-1023
 func encBinary(v []byte, b []byte) []byte {
 	var (
 		length  uint16
@@ -322,12 +328,12 @@ func encBinary(v []byte, b []byte) []byte {
 	)
 
 	if len(v) == 0 {
-		return encByte(b, BC_BINARY_DIRECT)
+		//return encByte(b, BC_BINARY_DIRECT)
+		return encByte(b, BC_NULL)
 	}
 
 	vLength = len(v)
 	for vLength > 0 {
-		// if vBuf.Len() > CHUNK_SIZE {
 		if vLength > CHUNK_SIZE {
 			length = CHUNK_SIZE
 			b = encByte(b, byte(BC_BINARY_CHUNK), byte(length>>8), byte(length))
@@ -405,34 +411,52 @@ func (e *Encoder) encUntypedMap(m map[interface{}]interface{}) error {
 	return nil
 }
 
-func buildMapKey(key reflect.Value, typ reflect.Type) interface{} {
-	switch typ.Kind() {
-	case reflect.String:
-		return key.String()
+func getMapKey(key reflect.Value, t reflect.Type) (interface{}, error) {
+	switch t.Kind() {
 	case reflect.Bool:
-		return key.Bool()
-	case reflect.Int:
-		return int32(key.Int())
+		return key.Bool(), nil
+
 	case reflect.Int8:
-		return int8(key.Int())
+		return int8(key.Int()), nil
 	case reflect.Int16:
+		return int16(key.Int()), nil
 	case reflect.Int32:
-		return int32(key.Int())
+		return int32(key.Int()), nil
+	case reflect.Int:
+		return int(key.Int()), nil
 	case reflect.Int64:
-		return key.Int()
+		return key.Int(), nil
+
 	case reflect.Uint8:
-		return byte(key.Uint())
-	case reflect.Uint, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-		return key.Uint()
+		return byte(key.Uint()), nil
+	case reflect.Uint16:
+		return uint16(key.Uint()), nil
+	case reflect.Uint32:
+		return uint32(key.Uint()), nil
+	case reflect.Uint:
+		return uint(key.Uint()), nil
+	case reflect.Uint64:
+		return key.Uint(), nil
+
+	case reflect.Float32:
+		return float32(key.Float()), nil
+	case reflect.Float64:
+		return float64(key.Float()), nil
+
+	case reflect.Uintptr:
+		return key.UnsafeAddr(), nil
+
+	case reflect.String:
+		return key.String(), nil
 	}
 
-	// return nil
-	return jerrors.Errorf("unsuport key kind %s", typ.Kind().String())
+	return nil, jerrors.Errorf("unsuport map key kind %s", t.Kind().String())
 }
 
 func (e *Encoder) encMap(m interface{}) error {
 	var (
 		err   error
+		k     interface{}
 		typ   reflect.Type
 		value reflect.Value
 		keys  []reflect.Value
@@ -446,9 +470,9 @@ func (e *Encoder) encMap(m interface{}) error {
 	}
 	e.buffer = encByte(e.buffer, BC_MAP_UNTYPED)
 	for i := 0; i < len(keys); i++ {
-		k := buildMapKey(keys[i], typ)
-		if k == nil {
-			return nil
+		k, err = getMapKey(keys[i], typ)
+		if err != nil {
+			return jerrors.Annotatef(err, "getMapKey(idx:%d, key:%+v)", i, keys[i])
 		}
 		if err = e.Encode(k); err != nil {
 			return nil
@@ -538,15 +562,15 @@ func (e *Encoder) encStruct(v POJO) error {
 		idx    int
 		num    int
 		err    error
-		clsDef classDef
+		clsDef classInfo
 	)
 
 	vv := reflect.ValueOf(v)
 
 	// write object definition
 	idx = -1
-	for i = range e.clsDefList {
-		if v.JavaClassName() == e.clsDefList[i].javaName {
+	for i = range e.classInfoList {
+		if v.JavaClassName() == e.classInfoList[i].javaName {
 			idx = i
 			break
 		}
@@ -557,8 +581,8 @@ func (e *Encoder) encStruct(v POJO) error {
 			idx = RegisterPOJO(v)
 		}
 		_, clsDef, _ = getStructDefByIndex(idx)
-		idx = len(e.clsDefList)
-		e.clsDefList = append(e.clsDefList, clsDef)
+		idx = len(e.classInfoList)
+		e.classInfoList = append(e.classInfoList, clsDef)
 		e.buffer = append(e.buffer, clsDef.buffer...)
 	}