+++ codec/hessian/change_log.md
@@ -11,6 +11,7 @@
 ---
 - 1 improvement: pkg/errors -> juju/errors
 - 2 bug fix:  ReflectResponse 中返回值为struct的时候对返回值（reflect.Value）转换为struct 对象
+- 3 improvement: rewrite decBinary
 
 ### 2018-04-26 ###
 ---